# backend/openapi.json
{
  "components": {
    "schemas": {
...
      "CampusItem": {
        "properties": {
          "code": {
            "description": "校区代码",
            "title": "Code",
            "type": "string"
          },
          "name": {
            "description": "校区名称",
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "code",
          "name"
        ],
        "title": "CampusItem",
        "type": "object"
      },
...
      "CourseOperationRequest": {
        "properties": {
          "course_id": {
            "description": "课程任务 ID",
            "title": "Course Id",
            "type": "string"
          },
          "method": {
            "default": "bx-b-b",
            "description": "选课方式代码",
            "title": "Method",
            "type": "string"
          }
        },
        "required": [
          "course_id"
        ],
        "title": "CourseOperationRequest",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "contact": {
      "name": "USTB Manager",
      "url": "https://github.com/yourusername/ustb-manager"
    },
    "description": "\n## USTB Manager - 北科大教务管理系统API\n\n这是一个用于访问北京科技大学教务系统的API服务，提供以下功能：\n\n### 🔐 认证模块 (Authentication)\n- **二维码登录**: 使用微信扫码快速登录\n- **短信验证码登录**: 通过手机号接收验证码登录\n- **Cookie登录**: 使用浏览器Cookie直接登录（高级用户）\n- **会话管理**: 自动管理用户会话，支持后端重启后恢复\n\n### 📊 成绩查询模块 (Grades)\n- **成绩列表**: 支持分页、筛选的成绩查询\n- **GPA计算**: 自动计算总GPA和加权GPA\n- **学生信息**: 获取学生基本信息和完整档案\n- **培养方案**: 查看专业培养方案和课程要求\n- **学业进度**: 查询必修课完成情况和学分统计\n\n### 🔒 安全特性\n- HttpOnly Cookie保护\n- Session自动过期和清理\n- Cookie有效性验证\n- 支持CORS跨域访问\n\n### 📝 使用说明\n1. 首先调用认证接口（QR/SMS/Cookie）完成登录\n2. 登录成功后会自动设置session cookie\n3. 后续请求会自动携带cookie进行认证\n4. 可以调用 `/api/auth/status` 检查登录状态\n5. 使用 `/api/auth/logout` 退出登录\n\n### 🌐 环境信息\n- 开发环境: http://localhost:8000\n- API前缀: `/api`\n- 文档地址: `/docs` (Swagger UI) 或 `/redoc` (ReDoc)\n\n### 📞 技术支持\n如有问题，请查看各接口的详细文档说明。\n    ",
    "license": {
      "name": "MIT License"
    },
    "title": "USTB Manager API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/courses/announcements": {
      "get": {
        "description": "## 业务说明\n查询选课公告信息，包含选课时间安排、注意事项等。\n\n## 参数说明\n- xn: 学年（可选，不传则使用当前选课学年）\n- xq: 学期（可选，不传则使用当前选课学期）\n\n## 返回数据\n选课公告列表",
        "operationId": "get_announcements_api_courses_announcements_get",
        "parameters": [
          {
            "description": "学年，如 2025-2026",
            "in": "query",
            "name": "xn",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "学年，如 2025-2026",
              "title": "Xn"
            }
          },
          {
            "description": "学期，如 2",
            "in": "query",
            "name": "xq",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "学期，如 2",
              "title": "Xq"
            }
          },
          {
            "in": "cookie",
            "name": "ustb_sid",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Ustb Sid"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {},
                  "title": "Response Get Announcements Api Courses Announcements Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "查询选课公告",
        "tags": [
          "courses"
        ]
      }
    },
    "/api/courses/available": {
      "get": {
        "description": "## 业务说明\n获取可选课程列表，支持按选课方式、学院、类别、校区筛选。\n\n## 参数说明\n- method: 选课方式代码\n  - `bx-b-b`: 必修课\n  - `mooc-b-b`: MOOC在线课程\n  - `sztzk`: 素质拓展课\n  - `zytzk`: 专业拓展课\n  - `ggkrw`: 公共课\n- college: 开课学院代码（通过 /courses/colleges 获取）\n- category: 课程类别代码（通过 /courses/categories 获取）\n- campus: 校区代码（通过 /courses/campuses 获取）\n- keyword: 课程名称关键字\n\n## 返回数据\n- courses: 可选课程列表\n- total: 课程总数\n- total_credits: 总学分\n- selection_method: 当前查询的选课方式",
        "operationId": "get_available_courses_api_courses_available_get",
        "parameters": [
          {
            "description": "选课学年，如 2025-2026",
            "in": "query",
            "name": "xn",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "选课学年，如 2025-2026",
              "title": "Xn"
            }
          },
          {
            "description": "选课学期，如 2",
            "in": "query",
            "name": "xq",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "选课学期，如 2",
              "title": "Xq"
            }
          },
          {
            "description": "选课方式: bx-b-b(必修), mooc-b-b(MOOC), sztzk(素质拓展), zytzk(专业拓展), ggkrw(公共课)",
            "in": "query",
            "name": "method",
            "required": false,
            "schema": {
              "default": "bx-b-b",
              "description": "选课方式: bx-b-b(必修), mooc-b-b(MOOC), sztzk(素质拓展), zytzk(专业拓展), ggkrw(公共课)",
              "title": "Method",
              "type": "string"
            }
          },
          {
            "description": "开课学院代码",
            "in": "query",
            "name": "college",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "开课学院代码",
              "title": "College"
            }
          },
          {
            "description": "课程类别代码",
            "in": "query",
            "name": "category",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "课程类别代码",
              "title": "Category"
            }
          },
          {
            "description": "校区代码",
            "in": "query",
            "name": "campus",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "校区代码",
              "title": "Campus"
            }
          },
          {
            "description": "课程关键字搜索",
            "in": "query",
            "name": "keyword",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "课程关键字搜索",
              "title": "Keyword"
            }
          },
          {
            "in": "cookie",
            "name": "ustb_sid",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Ustb Sid"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Get Available Courses Api Courses Available Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "获取可选课程列表",
        "tags": [
          "courses"
        ]
      }
    },
    "/api/courses/campuses": {
      "get": {
        "description": "## 业务说明\n获取所有校区的列表，用于筛选课程。\n\n## 返回数据\n校区列表，每项包含：\n- code: 校区代码\n- name: 校区名称",
        "operationId": "get_campuses_api_courses_campuses_get",
        "parameters": [
          {
            "in": "cookie",
            "name": "ustb_sid",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Ustb Sid"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/CampusItem"
                  },
                  "title": "Response Get Campuses Api Courses Campuses Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "获取校区列表",
        "tags": [
          "courses"
        ]
      }
    },
    "/api/courses/cart": {
      "get": {
        "description": "## 业务说明\n查询当前购物车中的课程列表。\n\n## 返回数据\n- `courses`: 购物车中的课程列表\n- `total`: 课程总数\n- `total_credits`: 总学分",
        "operationId": "get_cart_api_courses_cart_get",
        "parameters": [
          {
            "description": "选课方式代码",
            "in": "query",
            "name": "method",
            "required": false,
            "schema": {
              "default": "bx-b-b",
              "description": "选课方式代码",
              "title": "Method",
              "type": "string"
            }
          },
          {
            "in": "cookie",
            "name": "ustb_sid",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Ustb Sid"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SelectedCoursesResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "查询选课购物车",
        "tags": [
          "courses"
        ]
      }
    },
    "/api/courses/cart/add": {
      "post": {
        "description": "## 业务说明\n将一门可选课程添加到购物车（不直接选课）。\n\n## 使用场景\n- 意愿选课模式下先加入购物车\n- 需要比较多门课程后再确认选课",
        "operationId": "add_to_cart_api_courses_cart_add_post",
        "parameters": [
          {
            "in": "cookie",
            "name": "ustb_sid",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Ustb Sid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CourseOperationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CourseOperationResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "添加课程到购物车",
        "tags": [
          "courses"
        ]
      }
    },
    "/api/courses/cart/remove": {
      "post": {
        "description": "## 业务说明\n从购物车中移除一门或多门课程。\n\n## 参数说明\n- `course_ids`: 要移除的课程 ID 列表（支持批量操作）",
        "operationId": "remove_from_cart_api_courses_cart_remove_post",
        "parameters": [
          {
            "in": "cookie",
            "name": "ustb_sid",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Ustb Sid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CartRemoveRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CourseOperationResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "从购物车移除课程",
        "tags": [
          "courses"
        ]
      }
    },
    "/api/courses/cart/submit": {
      "post": {
        "description": "## 业务说明\n将购物车中的所有课程提交确认选课。\n\n## 注意事项\n- 购物车中的所有课程将被批量提交\n- 提交后无法撤回，需要通过退课操作",
        "operationId": "submit_cart_api_courses_cart_submit_post",
        "parameters": [
          {
            "description": "选课方式代码",
            "in": "query",
            "name": "method",
            "required": false,
            "schema": {
              "default": "bx-b-b",
              "description": "选课方式代码",
              "title": "Method",
              "type": "string"
            }
          },
          {
            "in": "cookie",
            "name": "ustb_sid",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Ustb Sid"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CourseOperationResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "提交购物车确认选课",
        "tags": [
          "courses"
        ]
      }
    },
    "/api/courses/categories": {
      "get": {
        "description": "## 业务说明\n获取所有课程类别的列表，用于筛选课程。\n\n## 返回数据\n类别列表，每项包含：\n- code: 类别代码\n- name: 类别名称",
        "operationId": "get_categories_api_courses_categories_get",
        "parameters": [
          {
            "in": "cookie",
            "name": "ustb_sid",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Ustb Sid"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/CategoryItem"
                  },
                  "title": "Response Get Categories Api Courses Categories Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "获取课程类别列表",
        "tags": [
          "courses"
        ]
      }
    },
    "/api/courses/check-conflict": {
      "post": {
        "description": "## 业务说明\n在选课或加入购物车前调用，检测目标课程与已选课程是否存在时间冲突。\n\n## 参数说明\n- `course_id`: 课程任务 ID，可通过 `/courses/available` 获取\n- `method`: 选课方式代码\n\n## 返回数据\n- `has_conflict`: 是否存在时间冲突\n- `message`: 冲突详情说明",
        "operationId": "check_time_conflict_api_courses_check_conflict_post",
        "parameters": [
          {
            "in": "cookie",
            "name": "ustb_sid",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Ustb Sid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConflictCheckRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConflictCheckResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "检测选课时间冲突",
        "tags": [
          "courses"
        ]
      }
    },
    "/api/courses/colleges": {
      "get": {
        "description": "## 业务说明\n获取所有开课学院的列表，用于筛选课程。\n\n## 返回数据\n学院列表，每项包含：\n- code: 学院代码\n- name: 学院名称",
        "operationId": "get_colleges_api_courses_colleges_get",
        "parameters": [
          {
            "in": "cookie",
            "name": "ustb_sid",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Ustb Sid"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/CollegeItem"
                  },
                  "title": "Response Get Colleges Api Courses Colleges Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "获取开课学院列表",
        "tags": [
          "courses"
        ]
      }
    },
...
  },
...
}